refactor(variableStore): replace IEnvFileManager with IVariableStore and update imports
