+++ src/envilder/domain/ports/IVariableStore.ts
